% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_campaign.R
\name{check_folder_size}
\alias{check_folder_size}
\title{Checks the folder size}
\usage{
check_folder_size(folder_path)
}
\arguments{
\item{folder_path}{\code{str} Path to the folder.}
}
\value{
\code{int} Size of the folder in GB.
}
\description{
Check the folder size in GB.
}
\keyword{internal}
