% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_new_version.R
\name{check_new_version}
\alias{check_new_version}
\title{Checks whether there is a new version of the app on GitHub}
\usage{
check_new_version(local_version = get_app_version())
}
\arguments{
\item{local_version}{\code{str} Local version number}
}
\value{
\code{str} Message if new version is available otherwise NULL.
}
\description{
Checks whether there is a new version of the app on GitHub
}
\keyword{internal}
