% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_masques.R
\name{compile_masques}
\alias{compile_masques}
\title{Compiles the zone de sante masques}
\usage{
compile_masques(campaign_name)
}
\arguments{
\item{campaign_name}{\code{str} Name of the campaign.}
}
\value{
\code{str} URL to the national compiled masque.
}
\description{
Compiles the zone de sante masques
}
\examples{
\dontrun{
compile_masques("CAMPAGNE_new_campaign") # must be a valid campaign folder
}
}
