% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_masques_antenne.R
\name{compile_masques_antenne}
\alias{compile_masques_antenne}
\title{Compile masque at the antenne level}
\usage{
compile_masques_antenne(campaign_name)
}
\arguments{
\item{campaign_name}{\code{str} Name of the campaign.}
}
\value{
NULL, invisibly.
}
\description{
Compile masque at the antenne level
}
\keyword{internal}
