% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_masques_national.R
\name{compile_masques_national}
\alias{compile_masques_national}
\title{Compiles the zone de sante masques}
\usage{
compile_masques_national(campaign_name)
}
\arguments{
\item{campaign_name}{\code{str} Name of the campaign.}
}
\value{
\code{str} URL to the national compiled masque.
}
\description{
Compared to the original version of the function, this masque uses the
IMPORTRANGE() function in Google Sheets to copy over the information. Rather
than having multiple imports and exports.
}
\examples{
\dontrun{
compile_masques_v2("CAMPAGNE_new_campaign") # must be a valid campaign folder
}
}
