% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_masques_province.R
\name{compile_masques_province}
\alias{compile_masques_province}
\title{Compile masque at the province level}
\usage{
compile_masques_province(campaign_name)
}
\arguments{
\item{campaign_name}{\code{str} Name of the campaign.}
}
\value{
NULL, invisibly.
}
\description{
Compile masque at the province level
}
\keyword{internal}
