% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_masques_national.R
\name{complete_compiled_masque}
\alias{complete_compiled_masque}
\title{Check if the masque was compiled completely}
\usage{
complete_compiled_masque(dribble)
}
\arguments{
\item{dribble}{\code{dribble} A dribble of a summary masque, ideally.}
}
\value{
\code{str} The URL of the summary masque or NA, if the masque compilation was unsuccessful.
}
\description{
The function basically checks if the number of rows in the first sheet is
equal to the last sheet. This is a basic check to see if the summary masque
was compiled successfully without any interruptions.
}
\keyword{internal}
