% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_masques_national.R
\name{copy_sheet_info_to_summary_masque}
\alias{copy_sheet_info_to_summary_masque}
\title{Copy data from data masques into the summary masque}
\usage{
copy_sheet_info_to_summary_masque(summary_masque, templates, sheet_name)
}
\arguments{
\item{summary_masque}{\code{dribble} The masque where the data should be copied to.}

\item{templates}{\code{dribble} A dribble of dribbles to copy.}

\item{sheet_name}{\code{str} Name of the sheet to copy from the templates}
}
\description{
The function compiles the structure of how the Google Sheet
function IMPORTFROM() and fills in the appropriate arguments of the function.
This is then appended to the summary masque.
}
\keyword{internal}
