% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_daily_campaign_progress_heatmap.R
\name{create_daily_campaign_progress_heatmap}
\alias{create_daily_campaign_progress_heatmap}
\title{Creates a heatmap showing campaign progress daily}
\usage{
create_daily_campaign_progress_heatmap(summary)
}
\arguments{
\item{summary}{\code{tibble} Summary table output of \code{\link[=get_campaign_progress]{get_campaign_progress()}}.}
}
\value{
\code{ggplot2} ggplot object.
}
\description{
Creates a heatmap showing campaign progress daily
}
\examples{
\dontrun{
ss_dribble <- googledrive::drive_get("spreadsheet_url")
summary <- get_campaign_progress(ss_dribble)
create_daily_campaign_progress_heatmap(summary)
}
}
