% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_masques_national.R
\name{create_masque_database}
\alias{create_masque_database}
\title{Create a summary masque}
\usage{
create_masque_database(campaign_name, folder, template_dribble, level)
}
\arguments{
\item{campaign_name}{\code{str} Campaign name}

\item{folder}{\code{dribble} Where the masque should be created}

\item{template_dribble}{\code{dribble} An example of masque to use as a template.
This is simply either a province or zone de sante masque (determined by level) so
that the function knows which columns to copy.}

\item{level}{\code{str} "national" or "province". What level to compile at.}
}
\value{
\code{dribble} Dribble to the newly created summary masque.
}
\description{
Creates a summary masque. At the national level, the function will compile from
the province level masques.
}
\keyword{internal}
