% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_init_campaign.R
\name{drive_cp_zs_template_parallel}
\alias{drive_cp_zs_template_parallel}
\title{Copy template folder to the ZS folder}
\usage{
drive_cp_zs_template_parallel(template_dribble, zs_drive_folders)
}
\arguments{
\item{template_dribble}{\code{dribble} A dribble object for the template sheet.}

\item{zs_drive_folders}{\code{dribble} A dribble of zones de sante folders.}
}
\value{
\code{dribble} A dribble containing the template files of a campaign.
}
\description{
Copy template folder to the ZS folder
}
\keyword{internal}
