% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_init_campaign.R
\name{drive_init_campaign}
\alias{drive_init_campaign}
\title{Initialize an SIA campaign}
\usage{
drive_init_campaign(
  start_date,
  end_date,
  campaign_name = Sys.Date(),
  prov_target = NULL,
  antenne_target = NULL,
  zs_target = NULL,
  gdb = NULL,
  zs_masque
)
}
\arguments{
\item{start_date}{\code{str} Start date of the campaign.}

\item{end_date}{\code{str} End date of the campaign.}

\item{campaign_name}{\code{str} The name of the SIA campaign. Defaults to the
date ran.}

\item{prov_target}{\code{str} A province or a vector of province names.}

\item{antenne_target}{\code{str} Antenne or a vector of antenne.}

\item{zs_target}{\code{str} A vector of Zone de Santes targeted for the campaign.}

\item{gdb}{\code{rda} rda file for the Geo database.}

\item{zs_masque}{\code{str} Dribble of the masque template.}
}
\value{
Success message
}
\description{
Creates the proper folder structure for a campaign.
}
\examples{
\dontrun{
init_campaign()
}
}
\keyword{internal}
