% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_init_campaign.R
\name{drive_mkdir_parallel}
\alias{drive_mkdir_parallel}
\title{Create Drive folders into another folder in parallel}
\usage{
drive_mkdir_parallel(names, target_folder)
}
\arguments{
\item{names}{\code{list} A list of folder names.}

\item{target_folder}{\code{dribble} Dribble of the target folder.}
}
\value{
\code{dribble} A dribble of the recently created folders.
}
\description{
Create Drive folders into another folder in parallel
}
\keyword{internal}
