% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{escape_non_ascii_only}
\alias{escape_non_ascii_only}
\title{Function to escape only non-ASCII characters that are non-valid R}
\usage{
escape_non_ascii_only(line)
}
\arguments{
\item{line}{\code{str} A string.}
}
\value{
\code{str} A string to escape non-ASCII characters
}
\description{
Function to escape only non-ASCII characters that are non-valid R
}
\keyword{internal}
