% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_drive_sheets.R
\name{find_drive_sheets}
\alias{find_drive_sheets}
\title{Obtain spreadsheets within a Drive folder}
\usage{
find_drive_sheets(folder_dribble)
}
\arguments{
\item{folder_dribble}{\code{dribble} Dribble containing the folder to search to.}
}
\value{
\code{dribble} Dribble containing the Google Sheets dribbles.
}
\description{
Obtains spreadsheets within a folder, which may or may not contain subfolders.
}
\examples{
\dontrun{
folder <- googledrive::drive_get("folder_id")
sheets <- find_drive_sheets(folder)
}
}
