% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/card_container.R
\name{fluent_card}
\alias{fluent_card}
\title{Fluent-style Card Container}
\usage{
fluent_card(...)
}
\arguments{
\item{...}{UI elements to include inside the card.}
}
\value{
A \code{shiny.tag} \code{div} containing the wrapped content.
}
\description{
Creates a simple card container with Fluent UI
}
