% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_masques_national.R
\name{gather_data_templates_from_folder}
\alias{gather_data_templates_from_folder}
\title{Gather masques from a folder}
\usage{
gather_data_templates_from_folder(folder, level = "national")
}
\arguments{
\item{folder}{\code{dribble} Dribble of the folder.}

\item{level}{\code{str} "national" or "province". National will look for province
level masques while province will look for zone de sante level masques.}
}
\value{
\code{dribble} A dribble with each row corresponding to a data masque.
}
\description{
Gathers the data templates (masques) from a drive folder.
}
\keyword{internal}
