% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geo_data}
\alias{geo_data}
\title{Geographic and population database}
\format{
\subsection{\code{geo_data}}{

A data frame with five columns:
\describe{
\item{provinces}{Country name}
\item{antennes}{Antenne}
\item{zones_de_sante}{Zones de Sante}
\item{aires_de_sante}{Aires de Sante}
\item{population_totale}{Total population of aires de sante}
}
}
}
\usage{
geo_data
}
\description{
The database containing all the geographic names in DRC, including total populations,
at the Aires de Sante level. This database is used primarily as the source for filling
the Excel template, as well as the various drop-down menus in the app.
}
\keyword{datasets}
