% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_app_version.R
\name{get_app_version}
\alias{get_app_version}
\title{Get the version of the application}
\usage{
get_app_version()
}
\value{
\code{str} Version number.
}
\description{
Get the version of the application
}
\keyword{internal}
