% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_campaign_progress.R
\name{get_campaign_progress}
\alias{get_campaign_progress}
\title{Monitor campaign quality}
\usage{
get_campaign_progress(dribble, sheets = 5:8)
}
\arguments{
\item{dribble}{\code{dribble} Dribble containing ZS dribbles related to a specific campaign.}

\item{sheets}{\code{int} Sheet number. Corresponds to Jour 1-4 for sheets 5-8, respectively.}
}
\value{
\code{tibble} Containing summary of daily coverage
}
\description{
Obtains the completeness of coverage for each day of the campaign.
}
\examples{
\dontrun{
dribble <- drive_get("zs_spreadsheet_url")
summary <- get_campaign_progress(dribble)
}
}
