% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_section_info.R
\name{get_section_info}
\alias{get_section_info}
\title{Get information on the completeness of a section}
\usage{
get_section_info(dribble, sheet = NULL, name = NULL)
}
\arguments{
\item{dribble}{\code{dribble} A dribble object identifying a specific Google Sheet.}

\item{sheet}{\code{str} or \code{int} Specify sheet ("tab") either by name or position.}

\item{name}{\code{str} Name of the section.}
}
\value{
\code{dribble} A dribble with section information on completeness.
}
\description{
Get information on the completeness of a section
}
\examples{
\dontrun{
ss <- googlesheets4::drive_get("example_sheet")
ss_summary <- get_section_info(ss)
}
}
