% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sheet_info.R
\name{get_sheet_info}
\alias{get_sheet_info}
\title{Obtain the information from a dribble tab or tabs}
\usage{
get_sheet_info(dribble, sheets = 1:8)
}
\arguments{
\item{dribble}{\code{dribble} Can be single or multi-row dribble containing
only dribbles of spreadsheets.}

\item{sheets}{\code{int} The tab to look into. Defaults to reading the first through
the eighth tab.}
}
\value{
\code{tibble} A tibble containing all the information in a sheet.
}
\description{
Collect information related to missingness of section for a particular tab.
}
\examples{
\dontrun{
get_sheet_info(dribble, 1)
}
}
