% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_campaign.R
\name{init_campaign}
\alias{init_campaign}
\title{Initialize an SIA campaign}
\usage{
init_campaign(
  start_date,
  end_date,
  campaign_name = Sys.Date(),
  campaign_folder = getwd(),
  prov_target = NULL,
  antenne_target = NULL,
  zs_target = NULL,
  gdb = NULL,
  zs_masque = system.file("extdata", "zs_masque_template.xlsx", package = "rdcAVS"),
  output_folder = campaign_folder
)
}
\arguments{
\item{start_date}{\code{str} Start date of the campaign.}

\item{end_date}{\code{str} End date of the campaign.}

\item{campaign_name}{\code{str} The name of the SIA campaign. Defaults to the
date ran.}

\item{campaign_folder}{\code{str} Path to the campaign folder. Defaults to \code{\link[=getwd]{getwd()}}.}

\item{prov_target}{\code{str} A province or a vector of province names.}

\item{antenne_target}{\code{str} Antenne or a vector of antenne.}

\item{zs_target}{\code{str} A vector of Zone de Santes targeted for the campaign.}

\item{gdb}{\code{rda} rda file for the Geo database.}

\item{zs_masque}{\code{str} Path to the masque template.}

\item{output_folder}{\code{str} Where to output the campaign.}
}
\value{
Success message
}
\description{
Creates the proper folder structure for a campaign.
}
\examples{
\dontrun{
init_campaign()
}
}
\keyword{internal}
