% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_ascii_char}
\alias{is_ascii_char}
\title{Function to safely check if a character is ASCII}
\usage{
is_ascii_char(char)
}
\arguments{
\item{char}{\code{chr}}
}
\value{
\code{bool} Whether it is an ASCII character or not
}
\description{
Function to safely check if a character is ASCII
}
\keyword{internal}
