% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_masques_national.R
\name{num_to_col}
\alias{num_to_col}
\title{Convert column number to column name.}
\usage{
num_to_col(n)
}
\arguments{
\item{n}{\code{int} Positive integer corresponding to column location. (1 = "A").}
}
\value{
\code{str} The column name in A1 format.
}
\description{
Convert column number to column name.
}
\keyword{internal}
