% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{perm_data}
\alias{perm_data}
\title{Permissions database}
\format{
\subsection{\code{perm_data}}{

A data frame with seven columns:
\describe{
\item{name}{Name of the person}
\item{email}{Email address, preferably a gmail email.}
\item{level}{Either "global", "province", "antenne", or "zone de sante"}
\item{role}{Currently supported are: "reader", "writer", "commenter".
Reader has read-only access. Writer has edit access. Commenter can only comment.}
\item{province}{Total population of aires de sante}
\item{antenne}{Antenne folder to grant permission to}
\item{zone_de_sante}{Zone de Sante folder to grant permissions to}
}
}
}
\usage{
perm_data
}
\description{
The database containing all the permissions required to be set in the Google Drive
campaign folder.
}
\keyword{datasets}
