% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{process_non_ascii_file}
\alias{process_non_ascii_file}
\title{Resolve non-ASCII character issues}
\usage{
process_non_ascii_file(file_path, output_path)
}
\arguments{
\item{file_path}{\code{str} Path to file where non-ASCII characters are located.}

\item{output_path}{\code{str} Output path}
}
\description{
Resolve non-ASCII character issues
}
\keyword{internal}
