% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdcAVS-package.R
\docType{package}
\name{rdcAVS-package}
\alias{rdcAVS}
\alias{rdcAVS-package}
\title{rdcAVS: SIA Campaign Builder for DRC}
\description{
This app assists in creating campaigns and setting permissions for campaigns in google drive.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://mcuadera.github.io/rdcAVS/}
}

}
\author{
\strong{Maintainer}: Mervin Keith Cuadera \email{xrg9@cdc.com} (\href{https://orcid.org/0000-0003-4898-2659}{ORCID})

Authors:
\itemize{
  \item Aureol-Le-Rocher Ngako Njiacheu \email{aureollerocher@gmail.com}
}

}
\keyword{internal}
