% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_permissions.R
\name{set_permissions}
\alias{set_permissions}
\title{Sets the permissions across dribbles based on permissions table}
\usage{
set_permissions(campaign_name, permissions_table, dribble_files)
}
\arguments{
\item{permissions_table}{\code{tibble} A tibble with permissions info. Required
columns are: \code{emailAddress} and \code{role}.}

\item{dribble}{\code{dribble} A dribble object to assign permissions to.}
}
\value{
\code{NULL} upon success.
}
\description{
Sets the permissions of a dribble based on the permissions table. If more
than one record is passed unto the \code{dribble}, all records will have the same
permissions as specified by the \code{permissions_table}.
}
\details{
Please see \href{https://googledrive.tidyverse.org/reference/drive_share.html}{drive_share()}
for a full explanation on valid roles.
}
\examples{
\dontrun{}
}
\keyword{internal}
