% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_permissions.R
\name{set_permissions_parallel}
\alias{set_permissions_parallel}
\title{Set dribble permissions in parallel}
\usage{
set_permissions_parallel(dribble_permissions)
}
\arguments{
\item{dribble_permissions}{\code{tibble} A permission table with associated dribbles.}
}
\value{
\code{NULL}
}
\description{
Sets permissions of dribbles in parallel
}
\keyword{internal}
