% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_fluent_task_button.R
\docType{data}
\name{task_button_css}
\alias{task_button_css}
\title{CSS for taskButton spinner animation}
\format{
An object of class \code{character} of length 1.
}
\usage{
task_button_css
}
\description{
CSS for taskButton spinner animation
}
\keyword{internal}
