% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_fluent_task_button.R
\docType{data}
\name{task_button_js}
\alias{task_button_js}
\title{JS handlers for taskButton busy/ready states}
\format{
An object of class \code{character} of length 1.
}
\usage{
task_button_js
}
\description{
JS handlers for taskButton busy/ready states
}
\keyword{internal}
