% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_campaign.R
\name{validate_geography}
\alias{validate_geography}
\title{Validates inputs against the geodatabase}
\usage{
validate_geography(gdb, geo_vector, spatial_scale)
}
\arguments{
\item{gdb}{\verb{parquet conn} A connection to the geodatabase.}

\item{geo_vector}{\code{str} A name or a vector of names.}

\item{spatial_scale}{\code{str} Spatial scale. Valid values are:
\itemize{
\item "prov": province
\item "zs": zone de sante
\item "as": aires de sante
}}
}
\value{
\code{NULL}
}
\description{
Validates inputs against the geodatabase
}
\keyword{internal}
