% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{campagneApp}
\alias{campagneApp}
\title{Main function}
\usage{
campagneApp(...)
}
\arguments{
\item{...}{Optional parameters. Practically speaking, not used.}
}
\value{
None.
}
\description{
This function gets called to run the app.
}
\examples{
\dontrun{
campagneApp()
}
}
