% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdcAVS-package.R
\docType{package}
\name{rdcAVS-package}
\alias{rdcAVS}
\alias{rdcAVS-package}
\title{rdcAVS: SIA Campaign Builder for DRC}
\description{
This app assists in creating campaigns and setting permissions for campaigns in google drive.
}
\author{
\strong{Maintainer}: Mervin Keith Cuadera \email{xrg9@cdc.com}

}
\keyword{internal}
